fix(dashboard,car): fix DashboardSection.Heading crashing under Server Components
DashboardSection.Heading is a sub-component attached to the exported function as a static property. React Server Components only create a client reference for a module's named exports at its 'use client' boundary, not for properties assigned onto an export afterward, so <DashboardSection.Heading> resolved to undefined wherever it rendered from a Server Component (the dashboard overview page and its insurance/technical-inspection sections). It only worked by accident on pages that were themselves already 'use client'. section.tsx also lacked 'use client' despite calling createContext, which failed the production build outright once a Server Component imported it.

Give SectionHeading its own 'use client' directive and import it by name at the broken call sites instead of through the DashboardSection.Heading property. Break the section.tsx <-> compounds/heading/heading.tsx circular import by moving the context and hook into a sibling section-context.ts.
