+++ src/module/car/presentation/ui/sections/insurance-expiration/insurance-expiration.tsx
@@ -1,6 +1,7 @@
 import { twMerge } from 'tailwind-merge';
 
 import { DateExpirationTable } from '@/car/presentation/ui/tables/date-expiration/date-expiration';
+import { SectionHeading } from '@/dashboard/ui/section/compounds/heading/heading';
 import { DashboardSection } from '@/dashboard/ui/section/section';
 
 type InsuranceExpirationSectionProps = {
@@ -12,9 +13,7 @@ export function InsuranceExpirationSection({
 }: InsuranceExpirationSectionProps) {
   return (
     <DashboardSection className={twMerge('lg:flex lg:flex-col', className)}>
-      <DashboardSection.Heading headingLevel="h2">
-        Insurance expiration
-      </DashboardSection.Heading>
+      <SectionHeading headingLevel="h2">Insurance expiration</SectionHeading>
       <DateExpirationTable dateColumn="insuranceExpiration" label="Insurance" />
     </DashboardSection>
   );
