+++ src/module/car/presentation/ui/sections/technical-inspection-expiration/technical-inspection-expiration.tsx
@@ -1,6 +1,7 @@
 import { twMerge } from 'tailwind-merge';
 
 import { DateExpirationTable } from '@/car/presentation/ui/tables/date-expiration/date-expiration';
+import { SectionHeading } from '@/dashboard/ui/section/compounds/heading/heading';
 import { DashboardSection } from '@/dashboard/ui/section/section';
 
 type TechnicalInspectionExpirationSectionProps = {
@@ -12,9 +13,9 @@ export function TechnicalInspectionExpirationSection({
 }: TechnicalInspectionExpirationSectionProps) {
   return (
     <DashboardSection className={twMerge('lg:flex lg:flex-col', className)}>
-      <DashboardSection.Heading headingLevel="h2">
+      <SectionHeading headingLevel="h2">
         Technical inspection expiration
-      </DashboardSection.Heading>
+      </SectionHeading>
       <div className="lg:flex lg:grow lg:flex-col lg:justify-center">
         <DateExpirationTable
           dateColumn="technicalInspectionExpiration"
