+++ src/module/dashboard/ui/section/section.tsx
@@ -1,8 +1,9 @@
+'use client';
+
 import type { ComponentProps } from 'react';
-import { createContext, useId } from 'react';
+import { useId } from 'react';
 import { twMerge } from 'tailwind-merge';
 
-import { useContextGuard } from '@/common/presentation/hook/use-context-guard';
 import type { SectionVariants } from '@/ui/variants/section';
 import { sectionVariants } from '@/ui/variants/section';
 
@@ -10,18 +11,7 @@ import { SectionControls } from './compounds/controls/controls';
 import { SectionHeading } from './compounds/heading/heading';
 import { SectionSubtext } from './compounds/subtext/subtext';
 import { SectionText } from './compounds/text/text';
-
-type DashboardSectionContextValue = { headingId: string } | null;
-
-const DashboardSectionContext =
-  createContext<DashboardSectionContextValue>(null);
-
-export function useDashboardSection() {
-  return useContextGuard({
-    context: DashboardSectionContext,
-    componentName: 'DashboardSection',
-  });
-}
+import { DashboardSectionContext } from './section-context';
 
 type DashboardSectionProps = ComponentProps<'section'> & {
   variant?: SectionVariants;
