+++ src/module/dashboard/ui/sections/overview/overview.tsx
@@ -2,6 +2,7 @@ import { TotalOwnershipsSection } from '@/car/ownership/presentation/ui/sections
 import { InsuranceExpirationSection } from '@/car/presentation/ui/sections/insurance-expiration/insurance-expiration';
 import { TechnicalInspectionExpirationSection } from '@/car/presentation/ui/sections/technical-inspection-expiration/technical-inspection-expiration';
 import { CostsSection } from '@/car/service-log/presentation/ui/sections/costs/costs';
+import { SectionHeading } from '@/dashboard/ui/section/compounds/heading/heading';
 import { DashboardSection } from '@/dashboard/ui/section/section';
 
 export function OverviewSection() {
@@ -10,9 +11,7 @@ export function OverviewSection() {
       className="lg:flex lg:flex-col lg:self-stretch"
       variant="raw"
     >
-      <DashboardSection.Heading headingLevel="h1">
-        Overview
-      </DashboardSection.Heading>
+      <SectionHeading headingLevel="h1">Overview</SectionHeading>
       <div className="lg:mx-auto lg:flex lg:w-full lg:max-w-7xl lg:grow lg:items-center lg:justify-center">
         <div className="flex flex-col gap-5 lg:grid lg:w-full lg:grid-cols-[auto_1fr_auto]">
           <TotalOwnershipsSection className="lg:col-span-1 lg:min-w-xs" />
